requirements.txt
psycopg2-binary==2.9.9
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.5.2
joblib==1.4.2
